pyproject.toml
[project]
name = "tp-6"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "blinker==1.9.0",
    "click==8.2.0",
    "colorama==0.4.6",
    "dotenv==0.9.9",
    "flask==3.1.0",
    "flask-jwt-extended==4.7.1",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "markupsafe==3.0.2",
    "mysql-connector-python==9.3.0",
    "pyjwt==2.10.1",
    "python-dotenv==1.1.0",
    "werkzeug==3.1.3",
]
